chatbot_v3.0/requirements.txt
streamlit
openai
sentence-transformers
pinecone-client
BeautifulSoup4
python-dotenv